Move HashTreeUtil to datastructures.
