+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/util/GenesisGenerator.java
@@ -40,7 +40,6 @@ import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.SSZTypes.SSZList;
 import tech.pegasys.teku.ssz.SSZTypes.SSZMutableList;
 import tech.pegasys.teku.util.config.Constants;
-import tech.pegasys.teku.util.hashtree.HashTreeUtil;
 
 public class GenesisGenerator {
