+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/util/Merkleizable.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.hashtree;
+package tech.pegasys.teku.datastructures.util;
 
 import org.apache.tuweni.bytes.Bytes32;
